requirements.txt
pandas 
requests
geopandas
jupyterlab
pygris
azure-storage-blob
us
pyyaml
pymongo
census
